First Draft
This uses A^{-1} to solve the equation A.x = b. "A inverse" is calculated by matrix decomposition using numpy.linalg.svd(A). 
This inverse may not be an actual inverse and A^{-1}.A ---> 1 (tends to).
A is m*n matrix where m = the total no. of elements in all the three sets A, B, C and n = the connected triples over the three sets.
b is a column matrix of m*1 which is the result when A is dotted with x. 
x is expected to be a column matrix of n*1 which gives the distinct triple set indexes, wherever the value is close to 1.